Add Hard Java Solution for 493. Reverse Pairs from LeetCode

+++ LeetCode/Patterns/Binary Search/493. Reverse Pairs/Solution.java
@@ -25,10 +25,10 @@ class Solution {
         int j = 0;
 
         while(i<a.length && j<b.length) {
-            if ((long) a[i] > 2L * b[j]){
+            if(a[i]>2*b[j]){
                 count += a.length - i;
                 j++;
-            } 
+            }
             else i++;
         }
